% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{data_tranform}
\alias{data_tranform}
\title{Data Transform}
\usage{
data_tranform(z0, type = "sku")
}
\arguments{
\item{z0}{dataset with historical sales}

\item{type}{type of inputed dataset. it's sku or fills detalization}
}
\value{
data frame with tidy dataset
}
\description{
This function transformating data to tidy data
}
