% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{days_koef}
\alias{days_koef}
\title{Week seasonality
This function analyze last 8 weeks sales and calculate
coefficients of day inside week for each group of goods}
\usage{
days_koef(z0)
}
\arguments{
\item{z0}{dataset with historical sales}
}
\value{
data frame with percentage of the day in week
}
\description{
Week seasonality
This function analyze last 8 weeks sales and calculate
coefficients of day inside week for each group of goods
}
