% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fills_koef.R
\name{fills_koef}
\alias{fills_koef}
\title{Fills part}
\usage{
fills_koef(z0)
}
\arguments{
\item{z0}{dataset with historical sales}
}
\value{
data frame with percentage of the day in week
}
\description{
This function analyze  sales and calculate
coefficients for each FillID
}
