% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Filter historical data}
\usage{
filt(z0, type = "both", type_f = "sku")
}
\arguments{
\item{z0}{dataset with historical sales}

\item{type}{type of filtration ("both","upper","lower")}

\item{type_f}{type of forecast ("sku" or "fills")}
}
\value{
data frame with clear data
}
\description{
This function filter historical data from upper values and replace that to mean
}
