% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtNA.R
\name{filtNA}
\alias{filtNA}
\title{Filter historical data from promo actions}
\usage{
filtNA(z0, type = "sku")
}
\arguments{
\item{z0}{dataset with historical sales}

\item{type}{type of forecast ("sku" or "fills")}
}
\value{
data frame with clear data
}
\description{
This function filter historical data from  days with promo actions
}
