% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_compl.R
\name{hist_compl}
\alias{hist_compl}
\title{Complete History}
\usage{
hist_compl(z0, type = "zero", transf = "sku")
}
\arguments{
\item{z0}{dataset with historical sales}

\item{type}{"zero" or "drop". Type of dropped rows}

\item{transf}{"sku" or "fills" for sku or fills detalization}
}
\value{
data frame with full historical dataset
}
\description{
This function make complete history when some data is dropped
}
