% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{is.nan
Function analog to is.na for Nan values}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{variable for diagnostic}
}
\value{

}
\description{
is.nan
Function analog to is.na for Nan values
}
