% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{my_abc}
\alias{my_abc}
\title{ABC analyses
This function make ABC analyses for each SKU}
\usage{
my_abc(dt, dl = c(0.5, 0.5))
}
\arguments{
\item{dt}{dataset with historical sales}
}
\value{
data frame with ABC group for each SKU
}
\description{
ABC analyses
This function make ABC analyses for each SKU
}
