% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{run_forecast}
\alias{run_forecast}
\title{Demand forecast}
\usage{
run_forecast(z0, catg, hist = T, comp = "zero", filt = "both", A = 0.9,
  B = 0.8, C = 0.7)
}
\arguments{
\item{z0}{dataset with historical sales}

\item{catg}{dataset with classificators of SKU}

\item{hist}{avaliable of hist_comp function (for "cast" of history)}

\item{comp}{type of transform ("zero","drop")}

\item{filt}{type of filtration ("both","upper","lower")}

\item{A}{.9 coefficients for ABC group}

\item{B}{.8 coefficients for ABC group}

\item{C}{.7 coefficients ABC group}
}
\value{
data frame with forecast
}
\description{
This function make data transfom, filtration and calculate forecast
}
