% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{Saf_Stock}
\alias{Saf_Stock}
\title{Safety Stock}
\usage{
Saf_Stock(z0, A = 0.9, B = 0.8, C = 0.7)
}
\arguments{
\item{z0}{dataset with historical sales}

\item{A}{coefficients for ABC group}

\item{B}{coefficients for ABC group}

\item{C}{coefficients ABC group}
}
\value{
data frame with ABC group for each SKU
}
\description{
This function make Safety Stock value for each SKU
}
\details{
\itemize{
\item abc
\item cs
\item nor_sales
\item nor_qnt
\item sls
\item sum_sls
\item top
\item sum_sal
}
}
