% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Filter historical data
This function filter historical data from upper values and replace that to mean}
\usage{
filt(z0)
}
\arguments{
\item{z0}{dataset with historical sales}
}
\value{
data frame with clear data
}
\description{
Filter historical data
This function filter historical data from upper values and replace that to mean
}
