% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtNA.R
\name{filtNA}
\alias{filtNA}
\title{Filter historical data from promo actions
This function filter historical data from  days with promo actions}
\usage{
filtNA(z0)
}
\arguments{
\item{z0}{dataset with historical sales}
}
\value{
data frame with clear data
}
\description{
Filter historical data from promo actions
This function filter historical data from  days with promo actions
}
