% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Demand forecast}
\usage{
forecast(y)
}
\arguments{
\item{y}{dataset with historical filtered sales}
}
\value{
data frame with forecast
}
\description{
This function calculate demand forecast for next periods
using 5 methods and choose minimal MAPE
}
